Get jurisdiction level from result URL
The data doesn't contain information about the jurisdiction for
the results, that is whether the URL represents state results or
county results.  However, we've found that result URLs are either
for a state or a county and that they have a distinct format.

Based on this, we can detect level from the URL so the user doesn't have
to specify it as an explicit command line argument.

Add a utility function to detect jurisdiction levels from a results URL
and use this function in the CLI commands.

Addresses #18
https://github.com/openelections/clarify/issues/18