Fix votes shown in output CSV
The logic that prepared a `Result` object for CSV output was incorrectly setting the votes to the total votes for each cadidate instead of the votes for each candidate in each jurisdiction.

Include the vote type in the CSV results.

Reorder CSV columns to match examples in http://docs.openelections.net/data-entry/

Add logging in the `Parser` class and an option in the `results` command to make it easier to identify problems with parsing or application logic.

Addresses #18
https://github.com/openelections/clarify/issues/18